Bug fixes and all functionality
All functionality update added
Minor bug fixes
