+++ packages/client/src/components/Functions/fetchTransaction.js
@@ -2,7 +2,8 @@ import axios from "axios";
 
 const fetchTransactions = async (userName) => {
     try {
-      const response = await axios.get(`https://financemvc.onrender.com/fetchTransactions/fetchG?userName=${userName}`);
+      // const response = await axios.get(`https://financemvc.onrender.com/fetchTransactions/fetchG?userName=${userName}`);
+      const response = await axios.get(`https://financemvc.onrender.com/transactionOperation/fetchG?userName=${userName}`);
       console.log("erert");
       // console.log(response.data.transactions);
       return response.data.transactions;
