+++ packages/server/src/routes/billReminderRoute.js
@@ -1,16 +1,15 @@
 const express =require('express');
-const saveBillReminderController=require('../controllers/saveBillReminderController.js');
-const fetchBillsController=require('../controllers/fetchBillsController.js');
-const deleteBillReminderController=require('../controllers/deleteBillReminderController.js'); 
-const fetchBillAlertController=require('../controllers/fetchBillAlertController.js');
+
+
+//Update
+const billCRUDoperations=require('../controllers/billCRUDoperations.js');
 //router object
 const router= express.Router();
 
 //LOGIN || POST
-router.post('/saveBill',saveBillReminderController);
-router.get('/fetchBills',fetchBillsController);
-router.delete('/deleteBill',deleteBillReminderController);
-router.get('/fetchBillAlert',fetchBillAlertController);
-// router.get('/fetchInActiveBills',deleteBillReminderController);
+router.post('/saveBill',billCRUDoperations.saveBillReminder);
+router.get('/fetchBills',billCRUDoperations.fetchBills);
+router.delete('/deleteBill',billCRUDoperations.deleteBillReminder);
+router.get('/fetchBillAlert',billCRUDoperations.fetchBillAlert);
 
 module.exports=router;
