--- packages/server/src/routes/deleteTransactionRoute.js
@@ -1,10 +1,0 @@
-const express =require('express');
-const deleteTransactionController=require('../controllers/deleteTransactionController.js');
-
-//router object
-const router= express.Router();
-
-//LOGIN || POST
-router.delete('/delete',deleteTransactionController);
-
-module.exports=router;
