--- packages/server/src/routes/fetchTransactionRoute.js
@@ -1,10 +1,0 @@
-const express =require('express');
-const {fetchTransactionController,fetchTransactionsForGraphsController}=require('../controllers/fetchTransactionController.js');
-//router object
-const router= express.Router();
-
-//LOGIN || POST
-router.get('/fetch',fetchTransactionController);
-router.get('/fetchG',fetchTransactionsForGraphsController);
-
-module.exports=router;
